server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
asyncpg==0.29.0
python-multipart==0.0.12
aiofiles==24.1.0
aioboto3==13.1.1
openai==1.51.0
langgraph==0.2.34
langgraph-checkpoint-postgres>=2.0.0
psycopg[binary,pool]==3.2.3
pypdf==5.0.1
pdfplumber==0.11.4
python-docx==1.1.2
pillow==10.4.0
pdf2image==1.17.0
tenacity==9.0.0
httpx==0.27.2
reportlab==4.2.5
pyjwt==2.9.0
